Make testing xdist safe
